refactor(infra): move proxy admin decoder logic
